finished JWT and OAUTH

+++ src/main/java/com/internProject/shortly/DAO/UrlRepository.java
@@ -5,8 +5,8 @@ import com.internProject.shortly.entity.Url;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface UrlRepository extends CrudRepository<Url, Integer> {
-
+public interface UrlRepository extends CrudRepository<Url, Long> {
 
     Url getByShortUrl(String shortUrl);
+
 }
